Initial commit.
The source code is updated up to 25th of August14. The code
is based on the aerofoil simulation code. It is capable of
runing tandem aerofoils,save pressure results multiple times
during simulation and velocity and density on the last
saving state. If the code crashes at least saves results
until the last saving state. The meshing tool has a polinomic
function that controls the horizontal interface so the mesh at the
leading edge of the elements has a low level of skewness.