Cl calculation changed.
The normals should point the inside of the aerofoil not the outsite.
Also for the computation of the Cl the lift force has to be divided by
the dynamic pressure times the surface. The dynamic pressure is computed
with the mach number given from the input, not from the moving frame
velocity, hence remains constant.