Change the scaling of the aerofoils at begining
The scaling of the aerofoils happened at the end at the same time they
were rotated. That means the points of the aerofoil were defined first
and then scaled, translated and rotated. Consequently the grid size of
the elements was not consistend along the span of the second element
when WLE were used. The scaling now happens before points are defined.