Postprocessing tool created and mod on Main3d
MainPost.f90 can:
1)Compute Average
2)Compute Cl
3)Write tecplot file

To be able to use this tool the Main3D.f90 needed new tools and
modifications:
Now data is saved in the data folder for use it after running time in
the postprocessing tool. The data is writen in 20 files, one per block,
which can be then read by any number of processors, independently of the
number of processors that produced it in first place.

Tools created can be found in rpt.f90
=>postwrite(nwrite)
=>postread(nread)
=>datawrite

The tools used to compute the Cl have been implemented in the Main3D
too. So the old subroutine clComp is not used anymore. New subroutine
clpost(ele,dir,var,post) is used now. This can be found in rpt.f90 too.