Major Update
Main3D:
1)The cl that goes to the output now is
based on the flow orientation wich is set manually.
2)The signal is written with more accuracy

rpt:
1)clpost sub has been modified and now computes both directions when
called.
2)A new sub has been created that computes tau_w in simmulation time.
This sub is called on the new clpost

SingleTurbulence3d:
1)spongego has been modified so the inflow turbulence can be
progressively switched off.  The parameters governing it are tgustd(time
when it starts switching off) and tguste(time that it will take to
switch off)

GridSingle:
1)smod parameter introduced, it determines how much times bigger should
the cell on vertical boundaries next to the aerofoil compared to the
aerofoil's most close cells
2)the naca function now also includes the aerofoil thickness as a
parameter.

Subroutineso:
1)movef sub has been changed so now it allows to progressivelly variate
the direction of the flow from 0 to alphamax. The variation starts at
talphas and goes on for talphar.