Added recovering saved data capability
The code can reuse the data stored in the misc folder that is finally
used to create the tecplot file. So now if the data is not deleted can
be used to write a file that includes data from different runs.

To do so the parameters nto and iwrec in the inputo file need to be
adjusted:

nto=0 =>delete files, next run will start from scratch
nto=1 =>do not delete files, from this run onwards results stored
nto=2 =>do not delete files, use previous files to create tecplot too
iwrec=Preavious record count, this needs to be set according to the last
simulation (Check Display before using it)