Move restart writing to a subroutine
Performance analysis tool showed that some time was lost in the region
where the restart file is written. It is due to the file being deleted
each time and then opened again.
1) Writing of the restart files was moved from Main3D to rpt
2) The file is just opened, closed and deleted once. It deletes the
preavious version when ndati=0, and then it opens it. The file remains
opened for the rest of the simulation until ndati=ndata, then it is
finally closed.